first successful auto-throttle flight code

+++ src/ECE477/inc/SparkFun_TMF8801_Arduino_Library.h
@@ -117,7 +117,7 @@ void init_TMF8801(TMF8801_t* dev);
 void read_distance(TMF8801_t* dev);
 
 void update_TMF8801(TMF8801_t* dev);
-uint16_t get_distance_TMF8801(TMF8801_t* dev);
+float get_distance_TMF8801(TMF8801_t* dev);
 
 //Initializes STM32F446 and TMF8801
 bool _TMF8801_init(TMF8801_t* dev);
